+++ CookMaster/app/src/main/java/com/example/cookmaster/model/IngredientRecipe.java
@@ -1,12 +1,12 @@
 package com.example.cookmaster.model;
 
 public class IngredientRecipe {
-    public IngredientRecipe (int ingredient_id, int recipe_id, int quantity){
+    public IngredientRecipe (long ingredient_id, long recipe_id, int quantity){
         this.ingredient_id = ingredient_id;
         this.recipe_id = recipe_id;
         this.quantity = quantity;
     }
-    public int ingredient_id;
-    public int recipe_id;
+    public long ingredient_id;
+    public long recipe_id;
     public int quantity;
 }
